fix: fix installation issues
This commit changes the default install mode from "Dialog" to "Global" in the `ConfigEntity.kt` file.

Additionally, it modifies the `InstallerActivity.kt` to prevent re-resolving the installer when the activity is re-created, for example, due to a configuration change. The installer resolution process will now only occur if it's a fresh launch for a new installation task.