feat(installer): Add option to set install requester
This commit introduces a new experimental feature that allows setting the `OriginatingUid` (install requester) for installation sessions.

This behavior, which mirrors the system's, can be enabled via a new "Set Install Requester" switch in the Lab settings. When enabled, the app will attempt to resolve the UID from the calling package name. If the UID cannot be retrieved, it will default to `null`.

Additionally, this commit includes several dependency updates:
- `composeBom` to `2025.12.00`
- `compose` to `1.10.0`
- `material3` to `1.5.0-alpha10`
- `activity-compose` to `1.12.1`
- `capsule` to `2.1.2`